+++ addons/binding/org.openhab.binding.roku/src/main/java/org/openhab/binding/roku/internal/RokuHandlerFactory.java
@@ -17,7 +17,7 @@ import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
-import org.openhab.binding.roku.handler.RokuHandler;
+import org.openhab.binding.roku.internal.handler.RokuHandler;
 
 /**
  * The {@link RokuHandlerFactory} is responsible for creating things and thing
