+++ addons/binding/org.openhab.binding.roku/src/main/java/org/openhab/binding/roku/internal/handler/RokuHandler.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.roku.handler;
+package org.openhab.binding.roku.internal.handler;
 
 import static org.openhab.binding.roku.RokuBindingConstants.*;
